expose full topology in service provider, compute prompt in internal state
